Fix outdated warning when `latestVersion` is empty string
The `latestVersion` property of `repositoryMeta` objects used to be either not present (`null`, `undefined`) or hold an actual value.

For components for which no latest version was found, the `latestVersion` field now holds an empty string. The `compareVersion` function used to compare versions did not take this case into consideration, causing components to be erroneously labeled as outdated.

Signed-off-by: nscuro <nscuro@protonmail.com>